Remove `Sync` from `RObject`
Instead, `RObjectShelter` has become `Sync` so you can send references to `RThreadSafeObject` to other threads.

Also reconfigured `RDataViewer::start()` to send the `data` to the thread in a thread safe object, which is likely how we will do it in the future when we avoid formatting the whole dataset up front.