Use a special left open range of `(]`
When determining `node_at_point()`, we want things like `match(asN|) to associate the `|` with the `asN` identifier, i.e. the cursor "belongs" to the `asN` node so we know to supply completions for that node in particular. This requires using a closed right end like `]`.

We also want `match(|)` to be completely unambiguous when determining the node position. i.e. is the `|` with the `(` or `)` node? To solve this, I've made the left side of the range open, meaning that the `|` goes with the `(`, which also makes sense, I think.