Implement debugging fallback when there are no srcrefs (#249)
* Implement debugging fallback when there are no srcrefs

* Switch to explicit typing

* Rename `sources` -> `fallback_sources`

* Add comment around `handle_stdout()`

* Respond with real errors in `handle_source()`

* Switch to internal `debugger_stack_info()` and new tooling

* One more refinement to use `RFunction`

* `lines_flatten()` -> `lines_join()`

* Switch to a `stop()` for this hopefully impossible case

* Also works for `parse(text = <text>, keep.source = TRUE)`

* `tryCatch()` the `parse()` call

This at least lets the `contents` flow through, so the editor opens even though we don't know where we are

* Add fallback behavior for known non parseable cases

* Don't use strings as the replacement

If both `<environment>` and a user defined string of `"<environment>"` happen to exist there, then replacing the `<environment>` with a string version of itself results in `""<environment>""`, i.e. two empty strings surrounding an unparseable `<environment>`.

Instead we insert an identifier that will parse but looks intentionally weird.