Send `ExecutionState::Starting` after our `Welcome` message
Doing it this way ensures that we've actually got a subscriber on the other end, and ensures that the first message the SUB receives is the `Welcome` message. I'm pretty sure this was just getting dropped most of the time before, but weirdly it would reliably come through on my local Windows machine, which was messing up `cargo nextest run` runs because we expected the `Welcome` message first.