Bump to r-lib/tree-sitter-r@cbfa8fbce269d6e1c46bddc5c4fcc4cdee6c3e9f
An empty `()`, or `()` with >1 expressions is now a syntax error, which is good, but does make it harder for us to give a precise error message. When we have a custom R parser that can report the exact problem in the parse result, we can do much better. For now it's a lot of headache and shooting a moving target.