Use a "function context" to create more responsive completions (#819)
* Use a "function context" to create more responsive completions

* Oops, I meant to keep this logging

* Consistency pass re: text edits

* Move `use` calls into `mod` blocks to avoid `#[cfg(test)]` proliferation (#826)

* Try to convince copilot to stop adding custom messages to assertions

* Handle usage that is technically a call, but still looks like a reference

Also modify the helper that post-processes `sort_text` to recognize when we've already applied an intentional prefix

* Add tooling around `sort_text` and work on the tests

* There's an existing helper that gets the job done

* Only mention filename

* Remove unnecessary `.clone()`

* Rename this node

Co-authored-by: Davis Vaughan <davis@rstudio.com>

* Explicitly handle case like `dplyr::@` + some renaming and refactoring

* Be more explicit about what sort of node we are prepared to handle

* Introduce `cursor_at_end` field into function context + more

The main substantive move here is adding a boolean `cursor_at_end` field into the function context. The new test re: an exact match for a 3-letter function like `dir()` is the real-world usage that made me realize we needed this.

Also implemented lots of feedback from @lionel-'s review at the same time.

* Fix this comment

* Prepare the way to use this test helper elsewhere and more flexibly

* Streamline namespace tests

`completions_from_namespace()` gets a new signature in this PR. Lets also streamline the tests with utilities similar to those in the function context tests.

---------

Co-authored-by: Davis Vaughan <davis@rstudio.com>