Prevent calling `readline()` or `menu()` from error handler
From R's perspective, the handler runs _after_ the error was emitted.
That's why the user is able to see the error message before the recover
prompt. From our perspective though, we have to run the handler _before_
emitting the error, because all executed code needs to nested in an
execution request so that we can properly match output to a prompt on
the frontend side. The Jupyter protocol does not really support orphan
side effects (streams, input requests).