Add a built-in for setting the PPM repo in a running session.
This commit adds a `.ps.set_ppm_repo()` built-in that works the same way
as the existing CLI option, but allows the repo to be set at runtime.

It is intended to support making the
`positron.r.packageManagerRepository` setting apply without needing to
restart/recreate existing R sessions.

Part of https://github.com/posit-dev/positron/issues/10965.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>